don't try to delete an already deleted group membership
A group membership could have been deleted before this pre-processing
so if the membership isn't found don't try to delete it

closes #6145

Reviewed-on: https://gerrit.instructure.com/6606
Tested-by: Hudson <hudson@instructure.com>

+++ app/models/enrollment.rb
@@ -193,7 +193,7 @@ class Enrollment < ActiveRecord::Base
       # we have just deleted the user's enrollment in the group's course.
       # remove the leaving user from the group to keep the group happy
       membership = group.group_memberships.find_by_user_id(self.user_id)
-      membership.destroy
+      membership.destroy if membership
     end
   end
   protected :audit_groups_for_deleted_enrollments
